% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{aggregateStats}
\alias{aggregateStats}
\title{Given one knockoff test stat per variable, return one per group.}
\usage{
aggregateStats(stats, groups, FUN = mean)
}
\description{
Given one knockoff test stat per variable, return one per group.
}
