% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate.getQvals}
\alias{calibrate.getQvals}
\title{Given knockoff statistics, return q-values.}
\usage{
calibrate.getQvals(W, offset = 1)
}
\arguments{
\item{W}{Knockoff statistics, symmetric under H0}

\item{offset}{Control modified (offset=0) vs traditional (offset=1) FDR. Same as knockoff::knockoff.threshold.}
}
\description{
modified from knockoff::knockoff.threshold
}
