% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check.calibration}
\alias{check.calibration}
\title{Given a set of knockoff stats and ground-truth nonzero entries, compute
q-values and compare them against actual false discoveries.}
\usage{
check.calibration(
  ground_truth,
  W,
  targeted_fdrs = (1:10)/10,
  plot_savepath = NULL,
  verbose = F
)
}
\description{
Given a set of knockoff stats and ground-truth nonzero entries, compute
q-values and compare them against actual false discoveries.
}
