% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{checkGroups}
\alias{checkGroups}
\title{Make sure variable groups match data matrix (or any other matrix with one variable per column)}
\usage{
checkGroups(X, groups)
}
\description{
Make sure variable groups match data matrix (or any other matrix with one variable per column)
}
