% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{chooseAdversarialY}
\alias{chooseAdversarialY}
\title{Pick P(Y|X) purposefully to reveal bad calibration.}
\usage{
chooseAdversarialY(X, kmeans_centers)
}
\arguments{
\item{X}{Covariates from which you want to select an active set, but you are
afraid your modeling assumptions might be wrong.}

\item{kmeans_centers}{The diagnostic looks for situations where the variance of
a given variable differs sharply across clusters. You have to manually
specify how many clusters}
}
\value{
A list with metadata and values of Y.
Each Y is generated from a univariate step function.
metadata includes the true active variable,
the cutoff used in the step function,
and the variable likely to be found as a false positive
if homoskedastic knockoffs are used.
}
\description{
Pick P(Y|X) purposefully to reveal bad calibration.
}
