% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dimensional.R
\name{createHighDimensionalKnockoffs}
\alias{createHighDimensionalKnockoffs}
\title{Deprecated. See create.highDimensionalKnockoffs}
\usage{
createHighDimensionalKnockoffs(...)
}
\description{
Deprecated. See create.highDimensionalKnockoffs
}
