% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookc.R
\name{do_pearson_screen}
\alias{do_pearson_screen}
\title{Helper function for simple marginal screening via the pearson correlation.}
\usage{
do_pearson_screen(X, ko, y)
}
\description{
Helper function for simple marginal screening via the pearson correlation.
}
