% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{findWorstY}
\alias{findWorstY}
\title{Given X and various simulated Y|X, find the worst-calibrated Y|X.}
\usage{
findWorstY(
  X,
  knockoffs,
  split = rep(c(F, T), length.out = nrow(X)),
  y,
  ground_truth,
  statistic = knockoff::stat.glmnet_lambdasmax,
  plot_savepath = NULL,
  ...
)
}
\arguments{
\item{X}{@param knockoffs A real dataset and a corresponding model-X knockoff realization.
To average out over multiple knockoff realizations, pass in a list of matrices. Any number is fine
if your computer can handle it. More is better since the final step is to select just a
single y (the worst one) and estimate the mis-calibration for it.}

\item{split}{to get an unbiased estimate of the calibration for the worst-calibrated P(Y|X),
half the data are used for choosing it and the other half for estimating its calibration.}

\item{y}{@param ground_truth Two lists of the same length, one containing y's that you simulated
and another containing the indices of the variables in X used to simulate them.}

\item{statistic}{Function used to compute variable importance, e.g. knockoff::stat.glmnet_lambdasmax.}

\item{plot_savepath}{Passed to check.calibration}

\item{...}{Passed to statistic, e.g. n_lambda=100.}
}
\description{
Given X and various simulated Y|X, find the worst-calibrated Y|X.
}
