% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookc.R
\name{formAllLooks}
\alias{formAllLooks}
\title{Given the low-rank representations, update knockoffs to omit each variable.}
\usage{
formAllLooks(knockoffs, vars_to_omit, updates, statistic = NULL)
}
\arguments{
\item{updates}{@param knockoffs @param vars_to_omit
Inputs should be from \code{loadCompactLooks} or from \code{generateLooks(..., output_type = 'knockoffs_compact'}.)
Those functions return a list with the same names as the necessary args.}

\item{statistic}{Optional but useful for memory efficiency: instead of returning all knockoffs, compute statistics using this function, and return those instead.}

\item{k}{variable to omit.}
}
\description{
Given the low-rank representations, update knockoffs to omit each variable.
}
