% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookc.R
\name{formLookParameters}
\alias{formLookParameters}
\title{Return mean and covariance used to sample a look. This is only used for checking the math.}
\usage{
formLookParameters(precomputed_quantities, vars_to_omit, updates, k)
}
\description{
Return mean and covariance used to sample a look. This is only used for checking the math.
}
