% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{generateGaussianMixtureLooks}
\alias{generateGaussianMixtureLooks}
\title{Gaussian mixture analog to GenerateLooks.}
\usage{
generateGaussianMixtureLooks(
  X,
  mus,
  Sigmas,
  posterior_probs,
  method = c("asdp", "sdp", "equi", "group"),
  groups = NULL,
  diag_s = NULL,
  vars_to_omit = 1:ncol(X),
  statistic = knockoff::stat.glmnet_coefdiff,
  output_type = c("knockoffs", "knockoffs_compact", "statistics", "parameters",
    "pearson"),
  ...
)
}
\description{
Gaussian mixture analog to GenerateLooks.
}
