% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookc.R
\name{generateLooks}
\alias{generateLooks}
\title{Generate and use leave-one-out knockoffs for each variable in X.}
\usage{
generateLooks(
  X,
  mu,
  Sigma,
  method = c("asdp", "sdp", "equi", "group"),
  groups = NULL,
  diag_s = NULL,
  vars_to_omit = 1:ncol(X),
  statistic = knockoff::stat.glmnet_coefdiff,
  output_type = c("knockoffs", "knockoffs_compact", "statistics", "parameters",
    "pearson"),
  ...
)
}
\arguments{
\item{X}{n-by-p matrix of original variables.}

\item{mu}{vector of length p, indicating the mean parameter of the Gaussian model for \eqn{X}.}

\item{Sigma}{p-by-p covariance matrix for the Gaussian model of \eqn{X}.}

\item{method}{either "equi", "sdp" or "asdp" (default: "asdp").
This determines the method that will be used to minimize the correlation between the original variables and the knockoffs.}

\item{diag_s}{Square PxP matrix or a vector of length P, containing the pre-computed matrix S where
\deqn{s_{jk} = cov(x_k, x_j) - cov(x_k, \tilde x_j)}.
This will be computed according to \code{method}, if not supplied.}

\item{statistic}{function used to assess variable importance. In addition to the options from
the \code{knockoff} package, you can enter the string \code{"pearson"}. This will take advantage
of the low-rank knockoff updates computed by generateLooks. This is only used if \code{output_type}
is "statistics".}

\item{output_type}{If "knockoffs" (default), result is a list of knockoff realizations. List is of the same length as vars_to_omit.
This can still be faster for generateLooks than for generateLooksSlow, as low-rank updates are used internally.
If "knockoffs_compact", a low-rank, low-memory update for each knockoff is returned, along with the precomputed
knockoffs that they modify. This option is not available for generateLooksSlow.
If "parameters", mean and covariance for each is returned.
This is for internal use.
If "statistics", knockoffs themselves are not returned; rather, test statistics are returned.
See parameter \code{statistic.}
This will err if \code{statistic} is set to NULL.
If "pearson", statistics are returned like with \code{output_type="statistic"}.
It's fast but rigid: the \code{statistic} arg is ignored and Pearson correlations are used.}

\item{...}{Passed to statistic}
}
\value{
See parameter \code{return_type}.
}
\description{
Generate and use leave-one-out knockoffs for each variable in X.
}
