% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookc.R
\name{getUpdateK}
\alias{getUpdateK}
\title{Extract one update from the low-rank representation, while handling a tricky case properly.}
\usage{
getUpdateK(k, vars_to_omit, updates)
}
\description{
Extract one update from the low-rank representation, while handling a tricky case properly.
}
