% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dimensional.R
\name{highDimensionalKnockoffs}
\alias{highDimensionalKnockoffs}
\title{Create Gaussian knockoffs with dense covariance in a large p, small n setting.}
\usage{
highDimensionalKnockoffs(
  X,
  rho = 0.9,
  lambda = 0.1,
  output_type = c("knockoffs", "parameters")
)
}
\arguments{
\item{X}{input features}

\item{rho}{tuning parameter with range 0<rho<1. Larger is expected to work better.}

\item{lambda}{degree of shrinkage: 0 returns the sample correlations and 1}

\item{output_type}{What to return: either knockoffs themselves or parameters
for downstream use in e.g. unit tests or leave-one-outs.
returns correlations of all 0.
You can set this for optimal MSE using corpcor::estimate_lambda.}
}
\description{
Create Gaussian knockoffs with dense covariance in a large p, small n setting.
}
