% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{KNNTest}
\alias{KNNTest}
\title{Run the "KNN diagnostics" from Section 5.1 (page 14) of the Romano et al. Deep Knockoffs paper.}
\usage{
KNNTest(X, X_k, n_neighbors = 20)
}
\description{
Run the "KNN diagnostics" from Section 5.1 (page 14) of the Romano et al. Deep Knockoffs paper.
}
