% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{knockoffQvals}
\alias{knockoffQvals}
\title{Given knockoff statistics, return q-values.}
\usage{
knockoffQvals(W, offset = 1)
}
\arguments{
\item{W}{Knockoff statistics, symmetric under H0}
}
\description{
modified from knockoff::knockoff.threshold
}
