% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{marginal_screen}
\alias{marginal_screen}
\title{A simple, fast test statistic, symmetric under the null, useful for testing knockoff software}
\usage{
marginal_screen(X, X_k, y)
}
\description{
A simple, fast test statistic, symmetric under the null, useful for testing knockoff software
}
