% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{removeKFromGroups}
\alias{removeKFromGroups}
\title{Remove one variable from a list of groups, then shift everything down.}
\usage{
removeKFromGroups(groups, k)
}
\description{
Remove one variable from a list of groups, then shift everything down.
}
