% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{safeCorrelation}
\alias{safeCorrelation}
\title{Compute correlations, but constant input gives 0 instead of causing a panic}
\usage{
safeCorrelation(...)
}
\description{
Compute correlations, but constant input gives 0 instead of causing a panic
}
