% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{saveload.formAllLooks}
\alias{saveload.formAllLooks}
\title{Given the low-rank representations, update knockoffs to omit each variable.}
\usage{
saveload.formAllLooks(
  knockoffs,
  vars_to_omit,
  updates,
  statistic = NULL,
  X = NULL,
  ...
)
}
\arguments{
\item{updates}{@param knockoffs @param vars_to_omit
Inputs should be previously saved by \code{saveCompactLooks} and read by \code{loadCompactLooks} or from \code{generateLooks(..., output_type = 'knockoffs_compact'}.)
Those functions return a list with the same names as the necessary args.}

\item{statistic}{Optional but perhaps useful for memory efficiency: instead of returning all knockoffs, compute statistics using this function, and return those instead.}

\item{k}{variable to omit.}
}
\description{
Given the low-rank representations, update knockoffs to omit each variable.
}
