% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{saveload.loadCompactLooks}
\alias{saveload.loadCompactLooks}
\title{Load result of saveCompactLooks}
\usage{
saveload.loadCompactLooks(savepath, filetype = c("csv", "h5"))
}
\description{
Load result of saveCompactLooks
}
