% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{saveload.saveCompactLooks}
\alias{saveload.saveCompactLooks}
\title{Save all the leave-one-out knockoffs to disk, e.g. for reading into Python.}
\usage{
saveload.saveCompactLooks(looks, savepath, filetype = c("csv", "h5"))
}
\description{
Save them in their low-rank representation to avoid wasting space.
}
