% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{saveLoad__formOneLook}
\alias{saveLoad__formOneLook}
\title{Given the low-rank representations, update knockoffs to omit one variable.}
\usage{
saveLoad__formOneLook(knockoffs, vars_to_omit, updates, k)
}
\arguments{
\item{updates}{@param knockoffs @param vars_to_omit
Inputs should be from \code{loadCompactLooks} or from \code{generateLooks(..., output_type = 'knockoffs_compact'}.)
Those functions return a list with the same names as the necessary args.}

\item{k}{variable to omit.}
}
\description{
Given the low-rank representations, update knockoffs to omit one variable.
}
