% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{saveLoad__loadCompactLooks}
\alias{saveLoad__loadCompactLooks}
\title{Load result of saveCompactLooks}
\usage{
saveLoad__loadCompactLooks(savepath, filetype = c("csv", "h5"))
}
\description{
Load result of saveCompactLooks
}
