% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{simulateY}
\alias{simulateY}
\title{Given X, simulate Y|X and check calibration.}
\usage{
simulateY(
  X,
  knockoffs,
  reserved_knockoffs = NULL,
  statistic = knockoff::stat.glmnet_lambdasmax,
  groups = seq(ncol(X)),
  active_set_size = 1,
  n_sim = 500,
  FUN = function(x) as.numeric(x > 0),
  plot_savepath = NULL,
  shuddup = F,
  kmeans_centers = kmeans_centers,
  ...
)
}
\arguments{
\item{X}{@param knockoffs A real dataset and a corresponding model-X knockoff realization.
To average out over multiple knockoff realizations, pass in a list of matrices. Any number is fine
if your computer can handle it; the code will cycle through if it's less than n_sim.}

\item{reserved_knockoffs}{Like knockoffs, but used only after the worst P(Y|X) has been
chosen from among those tested. Its miscalibration is re-estimated using new knockoffs
to avoid bias due to selective inference.}

\item{statistic}{Function used to compute variable importance, e.g. knockoff::stat.glmnet_lambdasmax.}

\item{groups}{Groups for composite hypothesis testing.}

\item{active_set_size}{how many grups to include in the active set for each simulation.}

\item{n_sim}{How many simulations to perform.}

\item{FUN}{How to simulate Y given some x's, e.g. function(x) all(x>0) + rnorm(1).
Should produce numeric output (not logical, even if you use a step function).
Pass the string "adversarial" or "diverse" for custom options designed to reveal
hidden harms of heteroskedasticity and other unexpected pitfalls.}

\item{plot_savepath}{Passed to check.calibration}

\item{kmeans_centers}{passed to chooseAdversarialY}

\item{...}{Passed to statistic, e.g. n_lambda=100.}
}
\description{
Given X, simulate Y|X and check calibration.
}
