% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{sqrt_inv}
\alias{sqrt_inv}
\title{Get a symmetric inverse square root of an spd matrix via eigendecomposition}
\usage{
sqrt_inv(X, power = -0.5)
}
\description{
Get a symmetric inverse square root of an spd matrix via eigendecomposition
}
