% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{stat.CCA}
\alias{stat.CCA}
\title{A knockoff statistic capable of handling multivariate Y.}
\usage{
stat.CCA(X, X_k, y)
}
\description{
For a single variable or knockoff x, the scheme is to regress x (as the target)
on Y (as the covariates) and compute cor(x, hat x) where hat x are the fitted values.
}
