% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{stat.marginalScreen}
\alias{stat.marginalScreen}
\title{A simple, fast test statistic, symmetric under the null, useful for testing knockoff software}
\usage{
stat.marginalScreen(X, X_k, y)
}
\description{
A simple, fast test statistic, symmetric under the null, useful for testing knockoff software
}
