% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookc.R
\name{updateKnockoffs}
\alias{updateKnockoffs}
\title{Given the low-rank updates from generateLooks, update knockoffs to omit variable k.}
\usage{
updateKnockoffs(knockoffs, updates, k)
}
\arguments{
\item{knockoffs}{Knockoffs computed with variable k included.}

\item{updates}{updates from generateLooks with output_type = "knockoffs_compact"}

\item{k}{variable to omit.}
}
\description{
Given the low-rank updates from generateLooks, update knockoffs to omit variable k.
}
