% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookc.R
\name{computeGaussianKnockoffs}
\alias{computeGaussianKnockoffs}
\title{Compute S, mean, covariance, and realizations of Gaussian knockoffs.}
\usage{
computeGaussianKnockoffs(
  X,
  mu,
  Sigma,
  groups = NULL,
  method = c("asdp", "sdp", "equi", "group"),
  diag_s = NULL,
  output_type = c("knockoffs", "knockoffs_compact", "statistics", "parameters")
)
}
\description{
This function is copied and lightly modified from the function \code{create.gaussian}
in Matteo Sesia's R package 'knockoff', licensed
under GPLv3 and published at https://github.com/msesia/knockoff-filter/tree/master/R .
I needed to modify it to return some internals and handle groups of variables.
}
\details{
Results of "output_type" are downright mislabeled -- that's just a holdover from generateLooks and it's not meant for public use.
}
