% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{computeGaussianMixtureKnockoffs}
\alias{computeGaussianMixtureKnockoffs}
\title{Deprecated. See create__gaussianMixtureKnockoffs}
\usage{
computeGaussianMixtureKnockoffs(...)
}
\description{
Deprecated. See create__gaussianMixtureKnockoffs
}
