% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{computeGaussianMixtureKnockoffs}
\alias{computeGaussianMixtureKnockoffs}
\title{Gaussian mixture analog to computeGaussianKnockoffs.}
\usage{
computeGaussianMixtureKnockoffs(
  X,
  mus,
  sigmas,
  posterior_probs,
  diag_s = NULL,
  output_type = c("knockoffs", "statistics", "parameters"),
  num_realizations = 1,
  seed = NULL,
  ...
)
}
\arguments{
\item{output_type}{If "knockoffs", a matrix matching X in shape.
If "parameters", a list with one element per cluster, containing parameters taken directly from the
"parameters" output of computeGaussianKnockoffs.}

\item{seed}{Meant for internal use only (in a unit test). This seed is set after sampling cluster assignments and before sampling knockoffs per cluster.}

\item{...}{passed to computeGaussianKnockoffs. See computeGaussianKnockoffs for parameter documentation.}
}
\description{
Gaussian mixture analog to computeGaussianKnockoffs.
}
