% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{simulateY}
\alias{simulateY}
\title{Given X, simulate Y|X and check calibration.}
\usage{
simulateY(
  X,
  knockoffs,
  statistic = marginalScreen,
  groups = seq(ncol(X)),
  active_set_size = 5,
  n_sim = 500,
  FUN = function(x) mean(x) + 0.1 * rnorm(1),
  plot_savepath = NULL,
  ...
)
}
\arguments{
\item{X}{@param knockoffs A real dataset and a corresponding model-X knockoff realization.
To average out over multiple knockoff realizations, pass in a list of matrices. Any number is fine
if your computer can handle it; the code will cycle through if it's less than n_sim.}

\item{statistic}{Function used to compute variable importance, e.g. knockoff::stat.glmnet_lambdasmax.}

\item{groups}{Groups for composite hypothesis testing.}

\item{active_set_size}{how many grups to include in the active set for each simulation.}

\item{n_sim}{How many simulations to perform.}

\item{FUN}{How to simulate Y given some x's, e.g. function(x) all(x>0) + rnorm(1)}

\item{plot_savepath}{Passed to check.calibration}

\item{...}{Passed to statistic, e.g. n_lambda=100.}
}
\description{
Given X, simulate Y|X and check calibration.
}
